Initial commit.
Basically, we ported a whole bunch of stuff over from
https://code.secondbit.org/api. But this iteration also includes some
helpers for Go clients decoding the errors returned. Basically, we
define a framework for turning a list of 0 or more errors in a response
body into a github.com/hashicorp/go-multierror.Error, filled with
user-defined errors. This makes parsing errors on the client-side much
more convenient.